separate phases for building AST and generating bindings
Separate binding generation into two phases, one for parsing the XML
source and extracting all definitions, and one for actually outputting
Rust code. The main reason this is necessary is that type names need to
have a prefix according to the extension in which they're defined, but a
given XML source file can refer to a type in any of the files in
imports, with or without an "extension:" prefix, so we need to be able
to look up names against the list of imports.

Implement "xidtype", "typedef", and "union" definitions in addition to
"struct"s.